+++ ros-examples/.catkin_workspace
@@ -1,0 +1,1 @@
+# This file currently only serves to mark the location of a catkin workspace for tool integration
