+++ ros-examples/src/aero_flight_modes/CMakeLists.txt
@@ -1,19 +1,19 @@
 cmake_minimum_required(VERSION 2.8.3)
-project(mavros_takeoff_land)
+project(aero_flight_modes)
 
-##Compile as C++ 11, supported in ROS Kinetic and newer
+## Compile as C++11, supported in ROS Kinetic and newer
 add_compile_options(-std=c++11 -g)
 
-##Find catkin macros and libraries
-##if COMPONENTS list likefind_package(catkin REQUIRED COMPONENTS xyz)
-##is used,also find other catkin packages
+## Find catkin macros and libraries
+## if COMPONENTS list like find_package(catkin REQUIRED COMPONENTS xyz)
+## is used, also find other catkin packages
 find_package(catkin REQUIRED COMPONENTS
   mavros_msgs
   roscpp
 )
 
 catkin_package(
-  CATKIN_DEPENDS roscpp mavros_msgs
+  CATKIN_DEPENDS mavros_msgs roscpp
 )
 
 ###########
@@ -21,11 +21,11 @@ catkin_package(
 ###########
 
 include_directories(
-${catkin_INCLUDE_DIRS}
-  )
+				${catkin_INCLUDE_DIRS}
+)
 
 add_executable(${PROJECT_NAME}_node
-  src/mavros_takeoff_land_node.cpp
+  src/aero_flight_modes_node.cpp
 )
 
 set_target_properties(${PROJECT_NAME}_node PROPERTIES OUTPUT_NAME node PREFIX "")
@@ -35,7 +35,7 @@ add_dependencies(${PROJECT_NAME}_node
   ${catkin_EXPORTED_TARGETS}
 )
 
-##Specify libraries to link a library or executable target against
+## Specify libraries to link a library or executable target against
 target_link_libraries(${PROJECT_NAME}_node
   ${catkin_LIBRARIES}
 )
@@ -44,8 +44,8 @@ target_link_libraries(${PROJECT_NAME}_node
 ## Install ##
 #############
 
-#all install targets should use catkin DESTINATION variables
+# all install targets should use catkin DESTINATION variables
 install(
- PROGRAMS
- DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
+  PROGRAMS
+  DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
 )
