+++ ros-examples/src/aero_fly_mission/CMakeLists.txt
@@ -1,19 +1,21 @@
 cmake_minimum_required(VERSION 2.8.3)
-project(mavros_telemetry_modes)
+project(aero_fly_mission)
 
 ## Compile as C++11, supported in ROS Kinetic and newer
-add_compile_options(-std=c++11 -g)
+add_compile_options(-std=c++11)
 
 ## Find catkin macros and libraries
 ## if COMPONENTS list like find_package(catkin REQUIRED COMPONENTS xyz)
 ## is used, also find other catkin packages
 find_package(catkin REQUIRED COMPONENTS
+  geometry_msgs
   mavros_msgs
+  sensor_msgs
   roscpp
 )
 
 catkin_package(
-  CATKIN_DEPENDS mavros_msgs roscpp
+   CATKIN_DEPENDS geometry_msgs mavros_msgs sensor_msgs roscpp
 )
 
 ###########
@@ -21,19 +23,14 @@ catkin_package(
 ###########
 
 include_directories(
-				${catkin_INCLUDE_DIRS}
+  ${catkin_INCLUDE_DIRS}
 )
 
 add_executable(${PROJECT_NAME}_node
-  src/mavros_telemetry_modes_node.cpp
+  src/aero_fly_mission_node.cpp
 )
 
 set_target_properties(${PROJECT_NAME}_node PROPERTIES OUTPUT_NAME node PREFIX "")
-
-add_dependencies(${PROJECT_NAME}_node
-  ${${PROJECT_NAME}_EXPORTED_TARGETS}
-  ${catkin_EXPORTED_TARGETS}
-)
 
 ## Specify libraries to link a library or executable target against
 target_link_libraries(${PROJECT_NAME}_node
@@ -44,8 +41,7 @@ target_link_libraries(${PROJECT_NAME}_node
 ## Install ##
 #############
 
-# all install targets should use catkin DESTINATION variables
 install(
   PROGRAMS
   DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
-)
+  )
