+++ ros-examples/src/aero_fly_mission/package.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <package>
-  <name>mavros_fly_mission</name>
-  <version>0.0.1</version>
-  <description>The mavros_fly_mission package</description>
+  <name>aero_fly_mission</name>
+  <version>0.0.0</version>
+  <description>  Demonstration of Flying missions of Intel Aero using MAVROS.</description>
   <maintainer email="shakthi.prashanth.m@intel.com">Shakthi Prashanth M</maintainer>
   <license>BSD 3-Clause</license>
