+++ ros-examples/src/aero_takeoff_land/package.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0"?>
 <package>
-<name>mavros_takeoff_land</name>
+<name>aero_takeoff_land</name>
 <version>0.0.0</version>
-<description>The mavros_takeoff_land package</description>
+<description> Demonstration of takeoff and land on Intel Aero using MAVROS bridge.</description>
 
-<maintainer email="padmashree.mandri@intel.com">padmashree mandri</maintainer>
+<maintainer email="padmashree.mandri@intel.com">Padmashree Mandri</maintainer>
 
 <license> BSD 3-Clause </license>
 <buildtool_depend>catkin</buildtool_depend>
