--- ros-examples/src/mavros_fly_mission/CMakeLists.txt
@@ -1,47 +1,0 @@
-cmake_minimum_required(VERSION 2.8.3)
-project(mavros_fly_mission)
-
-## Compile as C++11, supported in ROS Kinetic and newer
-add_compile_options(-std=c++11)
-
-## Find catkin macros and libraries
-## if COMPONENTS list like find_package(catkin REQUIRED COMPONENTS xyz)
-## is used, also find other catkin packages
-find_package(catkin REQUIRED COMPONENTS
-  geometry_msgs
-  mavros_msgs
-  sensor_msgs
-  roscpp
-)
-
-catkin_package(
-   CATKIN_DEPENDS geometry_msgs mavros_msgs sensor_msgs roscpp
-)
-
-###########
-## Build ##
-###########
-
-include_directories(
-  ${catkin_INCLUDE_DIRS}
-)
-
-add_executable(${PROJECT_NAME}_node
-  src/mavros_fly_mission_node.cpp
-)
-
-set_target_properties(${PROJECT_NAME}_node PROPERTIES OUTPUT_NAME node PREFIX "")
-
-## Specify libraries to link a library or executable target against
-target_link_libraries(${PROJECT_NAME}_node
-  ${catkin_LIBRARIES}
-)
-
-#############
-## Install ##
-#############
-
-install(
-  PROGRAMS
-  DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
-  )
